creates a hook to fetch movies data from TMDB and dispatch the update action to store data in movie slice.
